@@ -70,7 +70,7 @@
     },
 
     "require": {
-        "php": "^7.4|^8.0",
+        "php": "^8.0",
         "ext-curl": "*",
         "ext-json": "*",
         "ext-openssl": "*"
